pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "webinfo2md"
version = "0.1.0"
description = "Fetch web content and turn it into structured markdown with an LLM pipeline."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Codex" }
]
dependencies = [
  "beautifulsoup4>=4.12.3",
  "click>=8.1.8",
  "httpx>=0.28.1",
  "markdownify>=1.1.0",
  "pydantic>=2.10.6",
  "PyYAML>=6.0.2",
  "python-dotenv>=1.0.1",
  "readability-lxml>=0.8.1",
  "rich>=13.9.4",
]

[project.optional-dependencies]
playwright = [
  "playwright>=1.51.0",
]
openai = [
  "openai>=1.66.3",
]
anthropic = [
  "anthropic>=0.49.0",
]
deepseek = [
  "openai>=1.66.3",
]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.25.3",
  "pyright>=1.1.397",
  "ruff>=0.11.0",
]

[project.scripts]
webinfo2md = "webinfo2md.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/webinfo2md"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"
